requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
black==24.10.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
cssbeautifier==1.15.1
Django==4.2.17
django-allauth==65.3.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.3
django-htmx==1.21.0
django-phonenumber-field==8.0.0
django-widget-tweaks==1.5.0
djlint==1.36.3
EditorConfig==0.17.0
factory_boy==3.3.1
Faker==33.1.0
idna==3.10
iniconfig==2.0.0
jsbeautifier==1.15.1
json5==0.10.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
ruff==0.8.3
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.8.2
